requirements.txt
discord.py==2.3.2
python-dotenv
sqlmodel
asyncpg
cryptography
steamcom>=1.2.23
stripe
